chore/#1 : 폴더링 구조 변경
