+++ app/src/main/java/com/chattymin/sopt_compose/feature/signup/SignUpPage.kt
@@ -1,4 +1,4 @@
-package com.chattymin.sopt_compose.ui.feature.main
+package com.chattymin.sopt_compose.feature.signup
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -9,11 +9,11 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
 
 @Composable
-fun MainPage(navController: NavController) {
+fun SignUpPage(navController: NavController) {
     Box(
         modifier = Modifier.fillMaxSize(),
         contentAlignment = Alignment.Center
     ) {
-        Text(text = "Main")
+        Text(text = "SignUp")
     }
 }
