--- app/src/main/java/com/chattymin/sopt_compose/ui/feature/signin/SignInPage.kt
@@ -1,19 +1,0 @@
-package com.chattymin.sopt_compose.ui.feature.signin
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.navigation.NavController
-
-@Composable
-fun SignInPage(navController: NavController) {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(text = "Sign In")
-    }
-}
